DSP: Move interpreter source files to their own directory
Keeps code organized similar to the DSP JIT.
