+++ Source/Core/Core/Core.vcxproj
@@ -61,17 +61,17 @@
     <ClCompile Include="DSP\DSPCore.cpp" />
     <ClCompile Include="DSP\DSPEmitter.cpp" />
     <ClCompile Include="DSP\DSPHWInterface.cpp" />
-    <ClCompile Include="DSP\DSPIntArithmetic.cpp" />
-    <ClCompile Include="DSP\DSPIntBranch.cpp" />
-    <ClCompile Include="DSP\DSPIntCCUtil.cpp" />
-    <ClCompile Include="DSP\DSPInterpreter.cpp" />
-    <ClCompile Include="DSP\DSPIntExtOps.cpp" />
-    <ClCompile Include="DSP\DSPIntLoadStore.cpp" />
-    <ClCompile Include="DSP\DSPIntMisc.cpp" />
-    <ClCompile Include="DSP\DSPIntMultiplier.cpp" />
     <ClCompile Include="DSP\DSPMemoryMap.cpp" />
     <ClCompile Include="DSP\DSPStacks.cpp" />
     <ClCompile Include="DSP\DSPTables.cpp" />
+    <ClCompile Include="DSP\Interpreter\DSPIntArithmetic.cpp" />
+    <ClCompile Include="DSP\Interpreter\DSPIntBranch.cpp" />
+    <ClCompile Include="DSP\Interpreter\DSPIntCCUtil.cpp" />
+    <ClCompile Include="DSP\Interpreter\DSPInterpreter.cpp" />
+    <ClCompile Include="DSP\Interpreter\DSPIntExtOps.cpp" />
+    <ClCompile Include="DSP\Interpreter\DSPIntLoadStore.cpp" />
+    <ClCompile Include="DSP\Interpreter\DSPIntMisc.cpp" />
+    <ClCompile Include="DSP\Interpreter\DSPIntMultiplier.cpp" />
     <ClCompile Include="DSP\Jit\DSPJitArithmetic.cpp" />
     <ClCompile Include="DSP\Jit\DSPJitBranch.cpp" />
     <ClCompile Include="DSP\Jit\DSPJitCCUtil.cpp" />
@@ -289,13 +289,13 @@
     <ClInclude Include="DSP\DSPEmitter.h" />
     <ClInclude Include="DSP\DSPHost.h" />
     <ClInclude Include="DSP\DSPHWInterface.h" />
-    <ClInclude Include="DSP\DSPIntCCUtil.h" />
-    <ClInclude Include="DSP\DSPInterpreter.h" />
-    <ClInclude Include="DSP\DSPIntExtOps.h" />
-    <ClInclude Include="DSP\DSPIntUtil.h" />
     <ClInclude Include="DSP\DSPMemoryMap.h" />
     <ClInclude Include="DSP\DSPStacks.h" />
     <ClInclude Include="DSP\DSPTables.h" />
+    <ClInclude Include="DSP\Interpreter\DSPIntCCUtil.h" />
+    <ClInclude Include="DSP\Interpreter\DSPInterpreter.h" />
+    <ClInclude Include="DSP\Interpreter\DSPIntExtOps.h" />
+    <ClInclude Include="DSP\Interpreter\DSPIntUtil.h" />
     <ClInclude Include="DSP\Jit\DSPJitRegCache.h" />
     <ClInclude Include="DSP\LabelMap.h" />
     <ClInclude Include="ec_wii.h" />
