+++ Source/Core/Core/DSP/DSPAssembler.cpp
@@ -3,8 +3,11 @@
 // Licensed under GPLv2+
 // Refer to the license.txt file included.
 
+#include "Core/DSP/DSPAssembler.h"
+
 #include <cstdio>
 #include <cstdlib>
+#include <cstring>
 #include <fstream>
 #include <iostream>
 #include <map>
@@ -13,9 +16,7 @@
 #include "Common/CommonTypes.h"
 #include "Common/FileUtil.h"
 
-#include "Core/DSP/DSPAssembler.h"
 #include "Core/DSP/DSPDisassembler.h"
-#include "Core/DSP/DSPInterpreter.h"
 #include "Core/DSP/DSPTables.h"
 
 static const char* err_string[] = {"",
