+++ Source/Core/Core/DSP/DSPEmitter.cpp
@@ -2,14 +2,20 @@
 // Licensed under GPLv2+
 // Refer to the license.txt file included.
 
+#include "Core/DSP/DSPEmitter.h"
+
 #include <cstring>
+
+#include "Common/Assert.h"
+#include "Common/BitSet.h"
+#include "Common/CommonTypes.h"
+#include "Common/Logging/Log.h"
 
 #include "Core/DSP/DSPAnalyzer.h"
 #include "Core/DSP/DSPCore.h"
-#include "Core/DSP/DSPEmitter.h"
 #include "Core/DSP/DSPHost.h"
-#include "Core/DSP/DSPInterpreter.h"
 #include "Core/DSP/DSPMemoryMap.h"
+#include "Core/DSP/Interpreter/DSPInterpreter.h"
 
 #define MAX_BLOCK_SIZE 250
 #define DSP_IDLE_SKIP_CYCLES 0x1000
