+++ Source/Core/Core/DSP/DSPTables.cpp
@@ -4,12 +4,13 @@
 
 // Additional copyrights go to Duddie (c) 2005 (duddie@walla.com)
 
+#include "Core/DSP/DSPTables.h"
+
 #include "Common/CommonTypes.h"
 
 #include "Core/DSP/DSPEmitter.h"
-#include "Core/DSP/DSPIntExtOps.h"
-#include "Core/DSP/DSPInterpreter.h"
-#include "Core/DSP/DSPTables.h"
+#include "Core/DSP/Interpreter/DSPIntExtOps.h"
+#include "Core/DSP/Interpreter/DSPInterpreter.h"
 
 // clang-format off
 const DSPOPCTemplate opcodes[] =
