+++ Source/Core/Core/DSP/Jit/DSPJitCCUtil.cpp
@@ -4,8 +4,8 @@
 
 // Additional copyrights go to Duddie and Tratax (c) 2004
 
+#include "Core/DSP/DSPCore.h"
 #include "Core/DSP/DSPEmitter.h"
-#include "Core/DSP/DSPIntUtil.h"  // Helper functions
 
 using namespace Gen;
