+++ Source/Core/Core/DSP/Jit/DSPJitLoadStore.cpp
@@ -4,11 +4,11 @@
 
 // Additional copyrights go to Duddie and Tratax (c) 2004
 
+#include "Common/CommonTypes.h"
+
 #include "Core/DSP/DSPEmitter.h"
-#include "Core/DSP/DSPIntCCUtil.h"
-#include "Core/DSP/DSPIntUtil.h"
-#include "Core/DSP/DSPInterpreter.h"
 #include "Core/DSP/DSPMemoryMap.h"
+#include "Core/DSP/Interpreter/DSPInterpreter.h"
 
 using namespace Gen;
