+++ Source/Core/Core/DSP/Jit/DSPJitRegCache.cpp
@@ -2,11 +2,15 @@
 // Licensed under GPLv2+
 // Refer to the license.txt file included.
 
+#include "Core/DSP/Jit/DSPJitRegCache.h"
+
 #include <cinttypes>
+
+#include "Common/Assert.h"
+#include "Common/Logging/Log.h"
 
 #include "Core/DSP/DSPEmitter.h"
 #include "Core/DSP/DSPMemoryMap.h"
-#include "Core/DSP/Jit/DSPJitRegCache.h"
 
 using namespace Gen;
