+++ Source/Core/Core/HW/DSPLLE/DSPLLE.cpp
@@ -2,11 +2,12 @@
 // Licensed under GPLv2+
 // Refer to the license.txt file included.
 
+#include "Core/HW/DSPLLE/DSPLLE.h"
+
 #include <mutex>
 #include <string>
 #include <thread>
 
-#include "Common/Atomic.h"
 #include "Common/ChunkFile.h"
 #include "Common/CommonPaths.h"
 #include "Common/CommonTypes.h"
@@ -19,14 +20,11 @@
 #include "Core/DSP/DSPCore.h"
 #include "Core/DSP/DSPHWInterface.h"
 #include "Core/DSP/DSPHost.h"
-#include "Core/DSP/DSPInterpreter.h"
 #include "Core/DSP/DSPTables.h"
-#include "Core/HW/DSPLLE/DSPLLE.h"
+#include "Core/DSP/Interpreter/DSPInterpreter.h"
 #include "Core/HW/DSPLLE/DSPLLEGlobals.h"
 #include "Core/HW/Memmap.h"
 #include "Core/Host.h"
-#include "Core/Movie.h"
-#include "Core/NetPlayProto.h"
 
 static Common::Event dspEvent;
 static Common::Event ppcEvent;
