+++ Source/Core/Core/HW/DSPLLE/DSPLLE.h
@@ -8,6 +8,7 @@
 #include <mutex>
 #include <thread>
 
+#include "Common/CommonTypes.h"
 #include "Common/Flag.h"
 #include "Core/DSPEmulator.h"
